Rename module spectoh5 -> convert. spectoh5 is now a deprecated alias
